OK review transition kit issues.
Our transition was firing the same Events.
Because the update is kinda hard to track.
So we remove the old update async and just use a plain coroutine.
Super safe...it should fix the issues with the spam analytics data.
Warning Open Circle transition will fail...If I use on any game I cannot remember.